internet issues fixed and the daily call is now its own thing wow tha tis col
